Slots as iota and checking for errors;
Object that is pickable is kind of Item that can be stored in Inventory. If it is equippable as well, it may be Equipped to specific Inventory/Equipemnt slot.
It's necessary to check for errors: if item is equippable, but has wrong (ie SlotNA) slot set, or if is not equippable, but has SlotWeapon set.